global-scaler/requirements.txt
pytest
pykka
gevent
autopep8
prometheus-api-client
prometheus_client
numpy
kubernetes
requests
retry
pyyaml